Use the 3rd line of BC9001 signs as hostname
AddressString can contain hostnames or address.
AddressString constructor can resolve address if requested.
Add some functions to check if an address is resovable name or is an address.
Propagate hostname in station events.
Clean the 3rd line of BC9001 if it contains "Station"
Make dynmap display hostnames (need to reset the dynmap layer)